Continue migrating storagefiller
